+++ app/src/main/java/io/github/sds100/keymapper/BaseMainActivity.kt
@@ -29,14 +29,12 @@ import com.anggrayudi.storage.extension.toDocumentFile
 import io.github.sds100.keymapper.Constants.PACKAGE_NAME
 import io.github.sds100.keymapper.compose.ComposeColors
 import io.github.sds100.keymapper.databinding.ActivityMainBinding
-import io.github.sds100.keymapper.home.HomeViewModel
 import io.github.sds100.keymapper.mappings.keymaps.trigger.RecordTriggerController
 import io.github.sds100.keymapper.system.accessibility.AccessibilityServiceAdapter
 import io.github.sds100.keymapper.system.files.FileUtils
 import io.github.sds100.keymapper.system.inputevents.MyMotionEvent
 import io.github.sds100.keymapper.system.permissions.AndroidPermissionAdapter
 import io.github.sds100.keymapper.system.permissions.RequestPermissionDelegate
-import io.github.sds100.keymapper.util.Inject
 import io.github.sds100.keymapper.util.launchRepeatOnLifecycle
 import io.github.sds100.keymapper.util.ui.showPopups
 import kotlinx.coroutines.Dispatchers
@@ -55,8 +53,8 @@ abstract class BaseMainActivity : AppCompatActivity() {
         const val ACTION_SHOW_ACCESSIBILITY_SETTINGS_NOT_FOUND_DIALOG =
             "$PACKAGE_NAME.ACTION_SHOW_ACCESSIBILITY_SETTINGS_NOT_FOUND_DIALOG"
 
-        const val ACTION_USE_ASSISTANT_TRIGGER =
-            "$PACKAGE_NAME.ACTION_USE_ASSISTANT_TRIGGER"
+        const val ACTION_USE_FLOATING_BUTTONS =
+            "$PACKAGE_NAME.ACTION_USE_FLOATING_BUTTONS"
 
         const val ACTION_SAVE_FILE = "$PACKAGE_NAME.ACTION_SAVE_FILE"
         const val EXTRA_FILE_URI = "$PACKAGE_NAME.EXTRA_FILE_URI"
@@ -72,10 +70,6 @@ abstract class BaseMainActivity : AppCompatActivity() {
 
     val viewModel by viewModels<ActivityViewModel> {
         ActivityViewModel.Factory(ServiceLocator.resourceProvider(this))
-    }
-
-    private val homeViewModel by viewModels<HomeViewModel> {
-        Inject.homeViewModel(this)
     }
 
     private val currentNightMode: Int
@@ -162,16 +156,6 @@ abstract class BaseMainActivity : AppCompatActivity() {
                     viewModel.onCantFindAccessibilitySettings()
                     viewModel.handledActivityLaunchIntent = true
                 }
-
-                ACTION_USE_ASSISTANT_TRIGGER -> {
-                    findNavController(R.id.container).navigate(
-                        NavAppDirections.actionToConfigKeymap(
-                            keymapUid = null,
-                            showAdvancedTriggers = true,
-                        ),
-                    )
-                    viewModel.handledActivityLaunchIntent = true
-                }
             }
         }
 
@@ -201,6 +185,8 @@ abstract class BaseMainActivity : AppCompatActivity() {
     }
 
     override fun onDestroy() {
+        UseCases.onboarding(this).shownAppIntro = true
+
         viewModel.previousNightMode = currentNightMode
         unregisterReceiver(broadcastReceiver)
         super.onDestroy()