+++ app/src/main/java/io/github/sds100/keymapper/system/camera/CameraFlashInfo.kt
@@ -1,0 +1,7 @@
+package io.github.sds100.keymapper.system.camera
+
+data class CameraFlashInfo(
+    val supportsVariableStrength: Boolean,
+    val defaultStrength: Int,
+    val maxStrength: Int,
+)